feat: Implement drag-and-drop functionality for dashboard issues and enhance today task management
- Added `useDashboardDnD` hook to manage drag-and-drop interactions for moving issues between the todo list and today's tasks.
- Introduced `useIssueData` hook for managing issue list state and API fetching for the dashboard.
- Created `useTodayTaskHandlers` hook to provide optimistic handlers for completing, adding, removing, reordering, and pinning tasks.
- Developed MUI theme component overrides for a consistent UI experience across the application.
